backend/requirements.txt
# FastAPI & server
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0

# Database (async)
sqlalchemy[asyncio]>=2.0.36,<3.0.0
asyncpg>=0.30.0,<1.0.0
alembic>=1.14.0,<2.0.0

# Config & validation
pydantic[email]>=2.10.0,<3.0.0
email-validator>=2.0.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0

# Security (foundation only)
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Logging
python-json-logger>=3.2.0,<4.0.0

# Image tagging (download + Pillow)
httpx>=0.27.0,<1.0.0
pillow>=10.4.0,<12.0.0

# Outfit ranking (train offline; optional at runtime)
scikit-learn>=1.5.0,<2.0.0
numpy>=1.26.0,<3.0.0
